add .exe files for windows
