+++ dispenser/configuration/SFTPManager.py
@@ -1,7 +1,14 @@
 import json
-import os
 import time
 import platform
+
+sys_name = platform.system()
+
+import os, sys
+import subprocess
+
+if sys_name == "Windows":
+    import win32console, win32con
 
 
 class SFTPManager(object):
@@ -10,27 +17,25 @@ class SFTPManager(object):
     linux_unmount_cmd = "fusermount -uz {dir}"
     linux_mkdir_cmd = "mkdir -p {dir}"
 
-    windows_mount_cmd = "use {local_dir} \\sshfs\{user}}@{ip}}\{remote_dir}}"
-    windows_unmount_cmd = "use {dir} /delete"
-    windows_mkdir_cmd = "if not exist {dir} mkdir {dir}"
+    # windows_mount_cmd = "net use {local_dir} \\\\sshfs\\{user}@{ip}\\{remote_dir}"
+    windows_mount_cmd = 'sshfs-win.exe cmd {user}@{ip}:{remote_dir} {local_dir} -o password_stdin'
+    # windows_unmount_cmd = "net use {dir} /delete"
+    windows_unmount_cmd = 'taskkill /IM "sshfs.exe" /F'
 
     settings_file = "sftp_config.json"
 
     def __init__(self):
-        self.ip = ""
+        self.ip = "192.168.1.69"
         self.user = "orangepi"
         self.password = "orangepi"
         self.remote_dir = "/home/orangepi/PB.CNTRLS.HW/dispenser/main/resources"
 
         self.load()
 
-        self._sys_name = platform.system()
-
-        if self._sys_name == "Windows":
-            self.local_dir = "X:"
+        if sys_name == "Windows":
+            self.local_dir = "S:"
         else:
             self.local_dir = "remote_resources"
-
 
     def _mkdir_linux(self):
         cmd = SFTPManager.linux_mkdir_cmd.format(dir=self.local_dir)
@@ -52,32 +57,46 @@ class SFTPManager(object):
         cmd = SFTPManager.windows_unmount_cmd.format(dir=self.local_dir)
         os.popen(cmd)
 
-    def _mkdir_windows(self):
-        cmd = SFTPManager.windows_mkdir_cmd.format(dir=self.local_dir)
-        os.popen(cmd)
-
     def _mount_windows(self):
-        cmd = SFTPManager.linux_mount_cmd.format(user=self.user, ip=self.ip,
-                                            remote_dir=self.remote_dir,
-                                            local_dir=self.local_dir)
-        self._mkdir_windows()
-        time.sleep(0.5)
-        p = os.popen(cmd, "w")
-        time.sleep(1)
-        p.write(self.password)
-        time.sleep(1)
-        p.write(self.password)
+        cmd = SFTPManager.windows_mount_cmd.format(user=self.user, ip=self.ip,
+                                                   remote_dir=self.remote_dir,
+                                                   local_dir=self.local_dir)
+
+        os.chdir("C:\\Program Files\\SSHFS-Win\\bin")  # https://stackoverflow.com/a/7130809
+
+        # https://stackoverflow.com/a/44056313
+        try:
+            win32console.AllocConsole()
+        except win32console.error as exc:
+            if exc.winerror != 5:
+                raise
+
+        stdin = win32console.GetStdHandle(win32console.STD_INPUT_HANDLE)
+
+        p = subprocess.Popen(["cmd.exe", "/C", cmd], stdout=subprocess.PIPE)  # https://stackoverflow.com/a/9535231
+        time.sleep(2)
+        for c in "{}\r".format(self.password):  # end by CR to send "RETURN"
+            ## write some records to the input queue
+            x = win32console.PyINPUT_RECORDType(win32console.KEY_EVENT)
+            x.Char = unicode(c)
+            x.KeyDown = True
+            x.RepeatCount = 1
+            x.VirtualKeyCode = 0x0
+            x.ControlKeyState = win32con.SHIFT_PRESSED
+            stdin.WriteConsoleInput([x])
+
+        p.wait()
 
     def unmount(self):
-        if self._sys_name == "Linux":
+        if sys_name == "Linux":
             self._unmount_linux()
-        elif self._sys_name == "Windows":
+        elif sys_name == "Windows":
             self._unmount_windows()
 
     def mount(self):
-        if self._sys_name == "Linux":
+        if sys_name == "Linux":
             self._mount_linux()
-        elif self._sys_name == "Windows":
+        elif sys_name == "Windows":
             self._mount_windows()
 
     def save(self):
